@@ -1,11 +1,12 @@
 ﻿using Ryujinx.HLE.HOS.Services.SurfaceFlinger;
+using System;
 
 namespace Ryujinx.HLE.Ui
 {
     /// <summary>
     /// Information about the indirect layer that is being drawn to.
     /// </summary>
-    class RenderingSurfaceInfo
+    class RenderingSurfaceInfo : IEquatable<RenderingSurfaceInfo>
     {
         public ColorFormat ColorFormat { get; }
         public uint Width { get; }
